[metasrv] refactor: move cluster management concept Node and Slot into metasrv/raft/types/
