+++ metasrv/src/raft/types/cluster.rs
@@ -1,0 +1,40 @@
+// Copyright 2021 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::fmt;
+
+use serde::Deserialize;
+use serde::Serialize;
+
+use crate::raft::types::NodeId;
+
+/// A slot is a virtual and intermediate allocation unit in a distributed storage.
+/// The key of an object is mapped to a slot by some hashing algo.
+/// A slot is assigned to several physical servers(normally 3 for durability).
+#[derive(Serialize, Deserialize, Debug, Default, Clone)]
+pub struct Slot {
+    pub node_ids: Vec<NodeId>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq)]
+pub struct Node {
+    pub name: String,
+    pub address: String,
+}
+
+impl fmt::Display for Node {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}={}", self.name, self.address)
+    }
+}
