+++ metasrv/src/raft/types/mod.rs
@@ -11,12 +11,16 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+mod cluster;
 mod raft_txid;
 mod raft_types;
 
 #[cfg(test)]
 mod raft_types_test;
 
+pub use cluster::Node;
+pub use cluster::Slot;
 pub use raft_txid::RaftTxId;
 pub use raft_types::LogId;
 pub use raft_types::LogIndex;
