+++ metasrv/src/sled_store/sled_key_space.rs
@@ -28,10 +28,10 @@ use sled::IVec;
 use crate::meta_service::LogEntry;
 use crate::raft::state::RaftStateKey;
 use crate::raft::state::RaftStateValue;
-use crate::raft::state_machine::Node;
 use crate::raft::state_machine::StateMachineMetaKey;
 use crate::raft::state_machine::StateMachineMetaValue;
 use crate::raft::types::LogIndex;
+use crate::raft::types::Node;
 use crate::raft::types::NodeId;
 use crate::sled_store::SeqNum;
 use crate::sled_store::SledOrderedSerde;
